+++ apis/cassandra/v1beta1/cassandradatacenter_types.go
@@ -139,7 +139,7 @@ type CassandraDatacenterSpec struct {
 	// the number of racks cannot easily be changed once a datacenter is deployed.
 	Racks []Rack `json:"racks,omitempty"`
 
-	// Describes the persistent storage request of each server node
+	// StorageConfig describes the persistent storage request of each server node
 	StorageConfig StorageConfig `json:"storageConfig"`
 
 	// A list of pod names that need to be replaced.
@@ -250,7 +250,7 @@ type DseWorkloads struct {
 	SearchEnabled    bool `json:"searchEnabled,omitempty"`
 }
 
-// AdditionalVolumes StorageConfig defines additional storage configurations
+// AdditionalVolumes defines additional storage configurations
 type AdditionalVolumes struct {
 	// Mount path into cassandra container
 	MountPath string `json:"mountPath"`